requirements.txt
dotenv
Pillow==6.2.2
env==0.1.0
googletrans==3.0.0
numpy==1.18.5
oauthlib==3.1.0
packaging==20.4
pandas==1.0.4
plotly==4.8.1
streamlit==0.61.0
tweepy==3.8.0
vaderSentiment==3.3.2
watchdog==0.10.2
python-dotenv==0.13.0
bokeh

